committing all changes
